requirements.txt
black
colorama
coverage
coveralls
flake8
hypothesis
mypy
pytest
pyyaml
setuptools
twine
wheel
